requirements.txt
aioredis==1.3.1
amqp==2.5.2
asgiref==3.2.7
async-timeout==3.0.1
attrs==19.3.0
autobahn==20.4.3
Automat==20.2.0
billiard==3.6.3.0
celery==4.4.2
cffi==1.14.0
channels==2.4.0
channels-redis==2.4.2
constantly==15.1.0
cryptography==2.9.2
daphne==2.5.0
Django==2.2.8
django-celery-results==1.2.1
django-ckeditor==5.9.0
django-extensions==2.2.9
django-js-asset==1.2.2
django-model-utils==4.0.0
djangorestframework==3.11.0
djangorestframework-simplejwt==4.4.0
entrypoints==0.3
flake8==3.7.9
hiredis==1.0.1
hyperlink==19.0.0
idna==2.9
importlib-metadata==1.6.0
incremental==17.5.0
kombu==4.6.8
mccabe==0.6.1
msgpack==0.6.2
psycopg2==2.8.4
psycopg2-binary==2.8.4
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.5.0
pycparser==2.20
pyflakes==2.1.1
PyHamcrest==2.0.2
PyJWT==1.7.1
pyOpenSSL==19.1.0
python-decouple==3.2
pytz==2020.1
redis==3.5.0
service-identity==18.1.0
six==1.14.0
sqlparse==0.3.1
Twisted==20.3.0
txaio==20.4.1
vine==1.3.0
zipp==3.1.0
zope.interface==5.1.0
